Move converter logic from jsonObject to persistence/local
